+++ neo/client_test.go
@@ -13,6 +13,7 @@ func TestClient(t *testing.T) {
 	t.Run("NewClient()", func(t *testing.T) {
 		client := neo.NewClient(nodeURI)
 		assert.Equal(t, nodeURI, client.NodeURI)
+		assert.IsType(t, neo.Client{}, client)
 	})
 
 	t.Run(".GetBestBlockHash()", func(t *testing.T) {
@@ -20,12 +21,8 @@ func TestClient(t *testing.T) {
 			client := neo.NewClient(nodeURI)
 
 			blockHash, err := client.GetBestBlockHash()
-			assert.Nil(t, err)
-			assert.NotEqual(t, "", blockHash)
-		})
-
-		t.Run("SadCase", func(t *testing.T) {
-			// TODO
+			assert.NoError(t, err)
+			assert.NotEmpty(t, blockHash)
 		})
 	})
 
@@ -36,18 +33,14 @@ func TestClient(t *testing.T) {
 			for _, testBlock := range testBlocks {
 				t.Run(testBlock.id, func(t *testing.T) {
 					block, err := client.GetBlockByHash(testBlock.hash)
-					assert.Nil(t, err)
+					assert.NoError(t, err)
 
 					assert.Equal(t, testBlock.hash, block.Hash)
 					assert.Equal(t, testBlock.index, block.Index)
 					assert.Equal(t, testBlock.merkleRoot, block.Merkleroot)
-					assert.Equal(t, testBlock.numberOfTransactions, len(block.Transactions))
+					assert.Len(t, block.Transactions, testBlock.numberOfTransactions)
 				})
 			}
-		})
-
-		t.Run("SadCase", func(t *testing.T) {
-			// TODO
 		})
 	})
 
@@ -58,18 +51,14 @@ func TestClient(t *testing.T) {
 			for _, testBlock := range testBlocks {
 				t.Run(testBlock.id, func(t *testing.T) {
 					block, err := client.GetBlockByIndex(testBlock.index)
-					assert.Nil(t, err)
+					assert.NoError(t, err)
 
 					assert.Equal(t, testBlock.index, block.Index)
 					assert.Equal(t, testBlock.hash, block.Hash)
 					assert.Equal(t, testBlock.merkleRoot, block.Merkleroot)
-					assert.Equal(t, testBlock.numberOfTransactions, len(block.Transactions))
+					assert.Len(t, block.Transactions, testBlock.numberOfTransactions)
 				})
 			}
-		})
-
-		t.Run("SadCase", func(t *testing.T) {
-			// TODO
 		})
 	})
 
@@ -78,12 +67,8 @@ func TestClient(t *testing.T) {
 			client := neo.NewClient(nodeURI)
 
 			blockCount, err := client.GetBlockCount()
-			assert.Nil(t, err)
-			assert.NotEqual(t, 0, blockCount)
-		})
-
-		t.Run("SadCase", func(t *testing.T) {
-			// TODO
+			assert.NoError(t, err)
+			assert.NotEmpty(t, blockCount)
 		})
 	})
 
@@ -94,13 +79,9 @@ func TestClient(t *testing.T) {
 			expectedBlockHash := "3f0b498c0d57f73c674a1e28045f5e9a0991f9dac214076fadb5e6bafd546170"
 
 			blockHash, err := client.GetBlockHash(blockIndex)
-			assert.Nil(t, err)
-			assert.NotEqual(t, "", blockHash)
+			assert.NoError(t, err)
+			assert.NotEmpty(t, blockHash)
 			assert.Equal(t, expectedBlockHash, blockHash)
-		})
-
-		t.Run("SadCase", func(t *testing.T) {
-			// TODO
 		})
 	})
 
@@ -109,12 +90,8 @@ func TestClient(t *testing.T) {
 			client := neo.NewClient(nodeURI)
 
 			blockCount, err := client.GetConnectionCount()
-			assert.Nil(t, err)
-			assert.NotEqual(t, 0, blockCount)
-		})
-
-		t.Run("SadCase", func(t *testing.T) {
-			// TODO
+			assert.NoError(t, err)
+			assert.NotEmpty(t, blockCount)
 		})
 	})
 
@@ -125,16 +102,12 @@ func TestClient(t *testing.T) {
 			for _, testTransaction := range testTransactions {
 				t.Run(testTransaction.id, func(t *testing.T) {
 					transaction, err := client.GetTransaction(testTransaction.hash)
-					assert.Nil(t, err)
+					assert.NoError(t, err)
 
 					assert.Equal(t, testTransaction.hash, transaction.ID)
 					assert.Equal(t, testTransaction.size, transaction.Size)
 				})
 			}
-		})
-
-		t.Run("SadCase", func(t *testing.T) {
-			// TODO
 		})
 	})
 
@@ -147,16 +120,12 @@ func TestClient(t *testing.T) {
 					transactionOutput, err := client.GetTransactionOutput(
 						testTransactionOutput.hash, testTransactionOutput.index,
 					)
-					assert.Nil(t, err)
+					assert.NoError(t, err)
 
 					assert.Equal(t, testTransactionOutput.asset, transactionOutput.Asset)
 					assert.Equal(t, testTransactionOutput.value, transactionOutput.Value)
 				})
 			}
-		})
-
-		t.Run("SadCase", func(t *testing.T) {
-			// TODO
 		})
 	})
 
@@ -165,11 +134,7 @@ func TestClient(t *testing.T) {
 			client := neo.NewClient(nodeURI)
 
 			_, err := client.GetUnconfirmedTransactions()
-			assert.Nil(t, err)
-		})
-
-		t.Run("SadCase", func(t *testing.T) {
-			// TODO
+			assert.NoError(t, err)
 		})
 	})
 
@@ -178,7 +143,7 @@ func TestClient(t *testing.T) {
 			client := neo.NewClient(nodeURI)
 
 			ok := client.Ping()
-			assert.Equal(t, true, ok)
+			assert.True(t, ok)
 		})
 
 		t.Run("SadCase", func(t *testing.T) {
@@ -201,7 +166,7 @@ func TestClient(t *testing.T) {
 					client := neo.NewClient(testCase.uri)
 
 					ok := client.Ping()
-					assert.Equal(t, false, ok)
+					assert.False(t, ok)
 				})
 			}
 		})