+++ utility/base_58_test.go
@@ -46,7 +46,7 @@ func TestBase58(t *testing.T) {
 			for _, testCase := range testCases {
 				t.Run(testCase.description, func(t *testing.T) {
 					result, err := base58.Decode(testCase.in)
-					assert.Nil(t, err)
+					assert.NoError(t, err)
 					assert.Equal(t, testCase.out, hex.EncodeToString(result))
 				})
 			}
@@ -68,7 +68,45 @@ func TestBase58(t *testing.T) {
 			for _, testCase := range testCases {
 				t.Run(testCase.description, func(t *testing.T) {
 					_, err := base58.Decode(testCase.in)
-					assert.NotNil(t, err)
+					assert.Error(t, err)
+				})
+			}
+		})
+	})
+
+	t.Run(".Encode()", func(t *testing.T) {
+		t.Run("HappyCase", func(t *testing.T) {
+			testCases := []struct {
+				description string
+				in          string
+				out         string
+			}{
+				{
+					description: "Empty",
+					in:          "",
+					out:         "",
+				},
+				{
+					description: "Short",
+					in:          "4e19",
+					out:         "6wi",
+				},
+				{
+					description: "Long",
+					in:          "25793686e9f25b6b",
+					out:         "7GYJp3ZThFG",
+				},
+			}
+
+			base58 := utility.NewBase58()
+
+			for _, testCase := range testCases {
+				t.Run(testCase.description, func(t *testing.T) {
+					bytes, err := hex.DecodeString(testCase.in)
+					assert.NoError(t, err)
+
+					result := base58.Encode(bytes)
+					assert.Equal(t, testCase.out, result)
 				})
 			}
 		})